Read and Write NN from parameter 1d array
Squeezes every layer with all its weights and biases into a 1d array. First stores the weights, then the biases, then moves onto the next layer, to give you a clue as to how it functions. The writing is pretty much the same as the reading, except it takes parameters and two lines of code are flipped, which is something ive never done. Usually stuff isnt so revercible.